Revert move. Add annotation to WindowResourceAccountingTest
